src/main/java/com/whereToWatch/WhereToWatch/secruity/MyUserDetails.java
package com.whereToWatch.WhereToWatch.secruity;

import com.whereToWatch.WhereToWatch.models.Role;
import com.whereToWatch.WhereToWatch.models.User;
import com.whereToWatch.WhereToWatch.repos.UserRepo;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import javax.validation.constraints.Null;
import java.util.Collections;
import java.util.List;


@Service
public class MyUserDetails implements UserDetailsService {

    @Autowired
    private UserRepo userRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        final User user = userRepository.findByUsername(username);

        if (user == null) {
            throw new UsernameNotFoundException("User '" + username + "' not found");
        }

        List<Role> roles = Collections.singletonList(Role.ROLE_CLIENT);
        return org.springframework.security.core.userdetails.User//
                .withUsername(username)//
                .password(user.getPassword())//
                .accountExpired(false)//
                .authorities(roles)//
                .accountLocked(false)//
                .credentialsExpired(false)//
                .disabled(false)//
                .build();
    }

}